+++ suggest/memory/src/test/java/com/spotify/heroic/suggest/memory/MemoryBackendIT.java
@@ -8,7 +8,4 @@ public class MemoryBackendIT extends AbstractSuggestBackendIT {
     protected SuggestModule setupModule() {
         return MemorySuggestModule.builder().build();
     }
-
-    @Override
-    public void writeDuplicatesReturnErrorInResponse() { }
 }
